fixed bug in deleteCourse that caused crash, ui and usability improvements, cleaned up project

+++ src/main/java/com/example/ischedule/Controller/HomeController.java
@@ -1,7 +1,10 @@
 package com.example.ischedule.Controller;
 
 import com.example.ischedule.Model.*;
-import com.example.ischedule.Service.*;
+import com.example.ischedule.Service.CourseService;
+import com.example.ischedule.Service.PreferencesService;
+import com.example.ischedule.Service.RoomService;
+import com.example.ischedule.Service.UserService;
 import com.example.ischedule.security.CustomUserDetails;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
@@ -13,6 +16,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.sql.Time;
 import java.time.LocalTime;
@@ -22,29 +26,27 @@ import java.util.Optional;
 
 @Controller
 public class HomeController {
-
     private final UserService userService;
     private final CourseService courseService;
     private final PreferencesService preferencesService;
     private final RoomService roomService;
-    private final ScheduleService scheduleService;
 
-    public HomeController(UserService userService, CourseService courseService, PreferencesService preferencesService, RoomService roomService, CourseController courseController, ScheduleService scheduleService) {
+    public HomeController(UserService userService, CourseService courseService, PreferencesService preferencesService, RoomService roomService) {
         this.userService = userService;
         this.courseService = courseService;
         this.preferencesService = preferencesService;
         this.roomService = roomService;
-        this.scheduleService = scheduleService;
     }
 
     @GetMapping("/home")
     public String home(Model model) {
+        // Get currently logged-in user and add it to the Model
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         String username = auth.getName();
         model.addAttribute("username", username);
         CustomUserDetails currentUser = (CustomUserDetails) auth.getPrincipal();
 
-        // Determine the user's role and set corresponding attributes in the model
+        // Determine the user's role and set attributes in the model
         boolean isAdmin = currentUser.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"));
         boolean isAssistant = currentUser.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ASSISTANT"));
         boolean isStudent = currentUser.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_STUDENT"));
@@ -52,8 +54,8 @@ public class HomeController {
         model.addAttribute("isAssistant", isAssistant);
         model.addAttribute("isStudent", isStudent);
 
+        // Get the enrolled courses for the student and add them to the model
         if (isStudent) {
-            // Get the enrolled courses for the student and add them to the model
             User currentUserObject = userService.getUserByUsername(username);
             List<Course> enrolledCourses = courseService.getEnrolledCourses(currentUserObject.getId());
             model.addAttribute("enrolledCourses", enrolledCourses);
@@ -61,35 +63,40 @@ public class HomeController {
             model.addAttribute("isStudent", false);
         }
 
-        // Fetch preferences only for assistant and admin roles
+        // Fetch preferences and rooms only for assistant and admin roles for setting preferences
         Preferences preferences = null;
         List<Room> rooms = null;
         if (isAdmin || isAssistant) {
             preferences = preferencesService.getPreferencesByUserId(currentUser.getUser().getId());
             rooms = roomService.getAllRooms();
-            // Get all users with "admin" or "assistant" role
             List<User> adminAndAssistantUsers = userService.getUsersByRoleIn(UserRole.ADMIN, UserRole.ASSISTANT);
             model.addAttribute("adminAndAssistantUsers", adminAndAssistantUsers);
         }
         model.addAttribute("preferences", preferences);
         model.addAttribute("rooms", rooms);
 
-        // TODO: Add other necessary data to the model
+        // Get all courses for Students and Admins
         List<Course> allCourses = courseService.getAllCourses();
         model.addAttribute("allCourses", allCourses);
-
         return "home";
     }
 
     @Transactional
     @PostMapping("/deleteCourse")
     public String deleteCourse(@RequestParam("courseId") int courseId) {
-        courseService.deleteCourse(courseId);
+        Optional<Course> course = courseService.getCourseById(courseId);
+        if (course.isPresent()) {
+            Course courseToDelete = course.get();
+            for (User user : courseToDelete.getEnrolledUsers()) {
+                user.getEnrolledCourses().remove(courseToDelete);
+            }
+            courseToDelete.getEnrolledUsers().clear();
+            courseService.deleteCourse(courseId);
+        }
         return "redirect:/home";
     }
 
-
-    @Transactional //make sure every step of the whole transaction process is successful, otherwise db remains unchanged
+    @Transactional // Make sure every step of the whole transaction process is successful, otherwise db remains unchanged
     @PostMapping("/addCourse")
     public String addCourse(@ModelAttribute Course course,
                             @RequestParam("instructorId") int instructorId,
@@ -97,17 +104,19 @@ public class HomeController {
                             @RequestParam("startTime") String startTimeStr,
                             @RequestParam("endTime") String endTimeStr,
                             @RequestParam("roomId") int roomId) {
+        // Get current user to set admin_id for the new course
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         User admin = userService.getUserByUsername(auth.getName());
+        course.setAdmin(admin);
+        // Get user_id from template to set Instructor of the new course
         User instructor = userService.getUserById(instructorId);
         course.setInstructor(instructor);
-        course.setAdmin(admin);
 
         CourseSchedule courseSchedule = new CourseSchedule();
         courseSchedule.setCourse(course);
         courseSchedule.setDayOfWeek(dayOfWeek);
 
-        // Convert the string values to LocalTime
+        // Convert the string values to LocalTime to prevent parsing error
         LocalTime startTime = LocalTime.parse(startTimeStr, DateTimeFormatter.ISO_LOCAL_TIME);
         LocalTime endTime = LocalTime.parse(endTimeStr, DateTimeFormatter.ISO_LOCAL_TIME);
         courseSchedule.setStartTime(Time.valueOf(startTime));
@@ -115,7 +124,6 @@ public class HomeController {
 
         Room room = roomService.getRoomById(roomId);
         courseSchedule.setRoom(room);
-
         course.setCourseSchedule(courseSchedule);
 
         courseService.saveCourse(course);
@@ -155,13 +163,14 @@ public class HomeController {
                                   @RequestParam("endTime") String endTimeStr,
                                   @RequestParam("room") int roomId,
                                   @RequestParam("dayOfWeek") DayOfWeek dayOfWeek,
-                                  Authentication authentication) {
+                                  Authentication authentication,
+                                  RedirectAttributes redirectAttributes) {
         String username = authentication.getName();
         User user = userService.getUserByUsername(username);
         Preferences preferences = preferencesService.getPreferencesByUserId(user.getId());
         Room preferredRoom = roomService.getRoomById(roomId);
 
-        // Convert the string values to LocalTime
+        // Convert the string values to LocalTime to prevent parsing error
         LocalTime startTime = LocalTime.parse(startTimeStr, DateTimeFormatter.ISO_LOCAL_TIME);
         LocalTime endTime = LocalTime.parse(endTimeStr, DateTimeFormatter.ISO_LOCAL_TIME);
 
@@ -171,8 +180,9 @@ public class HomeController {
         preferences.setPreferredRoom(preferredRoom);
         preferences.setPreferredDayOfWeek(dayOfWeek);
         preferencesService.savePreferences(preferences);
+
+        redirectAttributes.addFlashAttribute("successMessage", "Preferences saved successfully!");
         return "redirect:/home";
     }
-
 }
